refactor(workspace): move stray *_tests.rs under the module they test
Eleven test files sat beside the module they covered as `<topic>_tests.rs`
rather than under a `tests` module owned by it, so finding a module's tests
meant knowing which of two layouts its crate happened to use. Each moves to
the owned shape the rest of the workspace already uses — `foo/tests.rs`, with
multi-file suites as topic submodules under `foo/tests/` — taking the
crate-absolute import form the already-converted suites use.

`crates/stella-runtime/src/wrapper/mod.rs` becomes `wrapper.rs`, finishing the
named-file module conversion for production code.

`openai.rs`'s inline `mod tests` moves out to `openai/tests.rs` so the moved
suite has a parent to hang from. That drops the file from 1919 lines to 725 and
retires it from the god-file baseline, so AGENTS.md's table and the crate README
follow. The body is copied verbatim and reindented by rustfmt rather than
dedented by hand, which would have rewritten the raw-string fixtures.

Three ripples, each of which a gate catches:

- `provider_parity.rs`'s `include_str!` witness corpus is repointed at the moved
  files, and gains `openai/tests.rs` — the witnesses the `openai` row cites for
  its cache and reasoning postures now live there (invariant 8).
- `scripts/prose-baseline.txt` is keyed by (file, pattern), so the moved files
  are judged at zero allowance. Fifteen of their seventeen pre-existing
  constructions are reworded and their baseline entries deleted. The two that
  remain are `door.variant` field accesses — the field the test exists to pin —
  and carry to the new path at 2, down from 11.
- `scripts/file-size-baseline.txt` gains one line each for `driver/tests.rs` and
  `anthropic/tests.rs`, the `mod` declaration for their new child, and loses
  `openai.rs` entirely.

`name_rung_corpus.jsonl` names some of these paths and is left alone: it is a
frozen snapshot of the tree on its harvest day, which `search_recall.rs` says.

`stella-graph`'s `parse/added_language_tests.rs` is deferred to #5421 — `parse.rs`
has an inline `mod tests` occupying the name, and the nested declaration that
would be valid Rust is not one `check-module-reachability` can resolve.

Supersedes #5381, whose base predates the move of every crate under `crates/`.

Closes #5420
