+++ crates/stella-cli/src/agent/persistence/tests.rs
@@ -1,0 +1,6 @@
+//! Suites over [`super`]'s persistence path, one topic per file.
+
+mod pipeline_variant;
+mod receipt_write;
+mod stream;
+mod usage_recovery;
