+++ crates/stella-context/src/retrieval.rs
@@ -951,8 +951,4 @@ pub(crate) use ranking::{
 pub(crate) use ranking::budget_tokens_for_bytes;
 
 #[cfg(test)]
-mod evidence_tests;
-#[cfg(test)]
-mod recall_tests;
-#[cfg(test)]
 mod tests;
