+++ crates/stella-context/src/store/tests.rs
@@ -501,7 +501,7 @@ async fn a_deferred_memory_loses_the_last_slot_to_a_normal_one() {
     // precedence is only visible when the durable fact wins the slot on band,
     // not on similarity (see the doc comment). The evidence gate would refuse
     // that unmatched memory outright, so the tier property is measured with
-    // the gate off; admission itself is pinned in `retrieval::evidence_tests`.
+    // the gate off; admission itself is pinned in `retrieval::tests::evidence`.
     .with_tuning(crate::retrieval::RecallTuning {
         require_evidence: false,
         ..crate::retrieval::RecallTuning::default()
