+++ crates/stella-core/src/driver/tests.rs
@@ -2601,6 +2601,7 @@ mod model_fallback;
 mod output_budget;
 mod parked_wait;
 mod provider_outcomes;
+mod requery;
 mod steer_midturn;
 mod usage_anchor;
 mod usage_completeness;
