+++ crates/stella-core/src/subagent.rs
@@ -1135,6 +1135,4 @@ fn last_assistant_text(messages: &[CompletionMessage]) -> Option<&str> {
 }
 
 #[cfg(test)]
-mod fork_scope_tests;
-#[cfg(test)]
 pub(crate) mod tests;
