+++ crates/stella-model/src/credential/tests.rs
@@ -1,3 +1,5 @@
+mod secondary_credential;
+
 use super::*;
 
 fn temp_credentials_path(name: &str) -> PathBuf {
