+++ crates/stella-runtime/README.md
@@ -267,7 +267,7 @@ it crosses.
 | `src/parts.rs` | The individual construction steps: `build_provider`, `open_store`, `seed_calibration`, `budget_guard`. (The tool registry needs no step: `ToolRegistry::new` takes only the workspace root, so the builder calls it directly.) |
 | `src/session.rs` | The composite: `RuntimeBuilder` assembling a `SessionRuntime`. Overridable per-step as hosts diverge (today: `with_provider`). |
 | `src/error.rs` | `RuntimeError` — typed construction failures. Degradations are `Notice`s on the built runtime instead, per the warn-never-disable posture. |
-| `src/wrapper/mod.rs` | The `TurnWrapper` trait (`before_turn`/`after_turn`), plus `admissible` and the `AdmittedContribution` it answers with — the value-level restatement of the manifest's undeclared-role and mistyped-signal load-time rules, and the only value the dispatcher will apply. |
+| `src/wrapper.rs` | The `TurnWrapper` trait (`before_turn`/`after_turn`), plus `admissible` and the `AdmittedContribution` it answers with — the value-level restatement of the manifest's undeclared-role and mistyped-signal load-time rules, and the only value the dispatcher will apply. |
 | `src/wrapper/dispatch.rs` | `WrapperDispatch` — the host sequence (#3494): stage program → `before_turn` per stage → the host's `TurnDriver` → `after_turn` → `judge` → `again`. Owns the loop; the host owns the turn. |
 | `src/wrapper/subprocess.rs` | `SubprocessWrapper` — the transport CI exercises: spawns `[runtime].argv`, writes the request as one line of JSON on stdin, reads the response on stdout, and `refuses_env_name` — the default-deny env-allowlist check `stella-cli`'s consent prompt also renders. |
 | `src/wrapper/in_process.rs` | `InProcessWrapper`/`WrapperHandler` — the fast path §3 permits for Rust, over the identical owned request/response types the subprocess transport uses. |
