+++ crates/stella-runtime/src/wrapper.rs
